Improve row removal "animation"
I put "animation" in quotes because it currently just pauses, as it did
before. But now it at least pauses on the frame that shows the full
rows, so that if a user does a hard drop that leads to row removal, the
full rows are actually shown so the user gets a clearer understanding of
what's happening.

So the end result is better, but the code got uglier, more convoluted.
There must be a cleaner way to do this...